Puts skeleton cards in wallet screen
