+++ src/components/CardSkeleton/CardSkeleton.tsx
@@ -1,0 +1,5 @@
+import { Skeleton, SkeletonProps } from '@chakra-ui/react';
+
+export type CardSkeletonProps = SkeletonProps;
+
+export const CardSkeleton = ({ ...props }: CardSkeletonProps) => <Skeleton {...props} />;
