+++ src/components/index.ts
@@ -11,3 +11,4 @@ export * from './FormWrapper';
 export * from './filters';
 export * from './DynamicTable';
 export * from './PasswordInput';
+export * from './CardSkeleton';
