+++ src/modules/wallet/pages/WalletScreen/components/Faq/FaqSkeleton.tsx
@@ -1,0 +1,4 @@
+import { CardSkeleton } from '@/components';
+
+export const FaqSkeleton = () =>
+  Array.from({ length: 4 }).map((_, index) => <CardSkeleton key={index} w="100%" h="6rem" />);
