+++ src/modules/wallet/pages/WalletScreen/components/Faq/index.ts
@@ -1,0 +1,1 @@
+export * from './Faq';
